Silence -Wshorten-64-to-32 warning
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@45382 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/userport/userport_wic64.c
@@ -1244,7 +1244,7 @@ static void send_reply_revised(const uint8_t rcode, const char *msg, const uint8
         stage_dummy = 1;
         bsr = stage_retcode = 1;
         bsl = stage_length = 2 + offs;
-        bsd = stage_data = len;
+        bsd = stage_data = (int)len;
         handshake_flag2();
     } else {
         /* legacy protocol */
